Adding an alternative fix for the case sensitive file names
For some reason the auto fix mechanism in the `CMakeLists.txt` file
didn't work for me. Sometimes it never fixed the included and gfortran
complained that `common.for` is not found. Sometimes
`${CMAKE_CURRENT_BINARY_DIR}/FORTRAN_temp` was created and the fix
worked, gfortran did not complain anymore, but the linking failed with:

    Writing the .pyf file with numpy.f2py, please be patient...
    Traceback (most recent call last):
      File "/home/yanez/projects/refprop/REFPROP-cmake/externals/REFPROP-headers/generate_header.py", line 172, in <module>
        dd = generate_function_dict('REFPROP.pyf')
      File "/home/yanez/projects/refprop/REFPROP-cmake/externals/REFPROP-headers/generate_header.py", line 131, in generate_function_dict
        output = find_subroutine(lines, istart)
      File "/home/yanez/projects/refprop/REFPROP-cmake/externals/REFPROP-headers/generate_header.py", line 91, in find_subroutine
        raise ValueError(lines[istart + 1 + offset].strip())
    ValueError: real dimension((ncmax)) :: x0
    CMakeFiles/refprop.dir/build.make:519: recipe for target 'librefprop.so' failed

So I wrote a small bash script that creates lower cased symlinks of
`COMMON.FOR` and `COMTRN.FOR`.