Adding cmake cache variable REFPROP_SOURCE_AUTOFIX
The auto fix mechanism doesn't work properly for. In most cases
compiling the fortran code was not the issue, as
`${CMAKE_CURRENT_BINARY_DIR}/FORTRAN_temp` was used for the source.
Linking was the issue:

        Writing the .pyf file with numpy.f2py, please be patient...
        Traceback (most recent call last):
          File "/home/yanez/projects/refprop/REFPROP-cmake/externals/REFPROP-headers/generate_header.py", line 172, in <module>
            dd = generate_function_dict('REFPROP.pyf')
          File "/home/yanez/projects/refprop/REFPROP-cmake/externals/REFPROP-headers/generate_header.py", line 131, in generate_function_dict
            output = find_subroutine(lines, istart)
          File "/home/yanez/projects/refprop/REFPROP-cmake/externals/REFPROP-headers/generate_header.py", line 91, in find_subroutine
            raise ValueError(lines[istart + 1 + offset].strip())
        ValueError: real dimension((ncmax)) :: x0
        CMakeFiles/refprop.dir/build.make:519: recipe for target 'librefprop.so' failed

I don't understand why this happend and I didn't want to go through
another repo fixing things for me.

So the easisest way was to add `REFPROP_SOURCE_AUTOFIX`, which defaults to
TRUE.

With  `cmake -DREFPROP_SOURCE_AUTOFIX=OFF` the whole autofix mechanism
is disabled and you can fix the file name problem yourself, for example
using the `utils/alternative_case_sens_fix` script from this repository.